Elevate properties state and call
